+++ Core/models/Volume.cs
@@ -238,5 +238,12 @@ namespace Oci.CoreService.Models
         [JsonProperty(PropertyName = "autotunePolicies")]
         public System.Collections.Generic.List<AutotunePolicy> AutotunePolicies { get; set; }
         
+        /// <value>
+        /// Reservations-enabled is a boolean field that allows to enable PR (Persistent Reservation) on a volume.
+        /// 
+        /// </value>
+        [JsonProperty(PropertyName = "isReservationsEnabled")]
+        public System.Nullable<bool> IsReservationsEnabled { get; set; }
+        
     }
 }